refactor: ConnectLedger component
