+++ packages/frontend/src/components/navigation/Navigation.js
@@ -1,13 +1,8 @@
-import { getLocation } from 'connected-react-router';
 import React, { useState, useEffect, useCallback } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
 import styled from 'styled-components';
 
-import { actions as ledgerActions, selectLedgerConnectionAvailable, selectLedgerHasLedger } from '../../redux/slices/ledger';
 import DesktopContainer from './DesktopContainer';
 import MobileContainer from './MobileContainer';
-
-const { handleShowConnectModal } = ledgerActions;
 
 const Container = styled.div`
     &&& {
@@ -44,13 +39,7 @@ export default ({
     availableAccounts,
     account
 }) => {
-    const dispatch = useDispatch();
-
     const [menuOpen, setMenuOpen] = useState(false);
-
-    const ledgerConnectionAvailable = useSelector(selectLedgerConnectionAvailable);
-    const hasLedger = useSelector(selectLedgerHasLedger);
-    const { pathname } = useSelector(getLocation);
 
     useEffect(() => {
         if (menuOpen) {
@@ -94,10 +83,6 @@ export default ({
         setMenuOpen(false);
     }, []);
 
-    const connectLedger =  () => dispatch(handleShowConnectModal());
-
-    const showConnectLedgerButton = hasLedger || ['sign-in-ledger', 'setup-ledger'].includes(pathname.split('/')[1]);
-
     return (
         <Container id='nav-container' open={menuOpen}>
             <DesktopContainer
@@ -110,9 +95,6 @@ export default ({
                 refreshBalance={refreshBalance}
                 availableAccounts={availableAccounts}
                 account={account}
-                connectLedger={connectLedger}
-                ledgerConnectionAvailable={ledgerConnectionAvailable}
-                showConnectLedgerButton={showConnectLedgerButton}
             />
             <MobileContainer
                 menuOpen={menuOpen}
